Remove duplicate enum and macro about numberic_enum_mod (#11347)
* remove duplicated and useless `remove duplicated and useless`

close #11297


* remove duplicate macro rocksdb_enum_mod


* fix format problem


* resolve discussion


* fix test error


* move macro `numeric_enum_serializing_mod` to a proper location


* move `perf_level_serde` of mod config in engine trait to mod perf_context


* Add test about `rate-limiter-mode` and fix compile error in integration test
